Interpolation search

+++ 102-interpolation.c
@@ -35,11 +35,11 @@ int interpolation_search(int *array, size_t size, int value)
 
 		if (pos >= size)
 		{
-			printf("Value checked array[%zu] is out of range\n", pos);
+			printf("Value checked array[%lu] is out of range\n", (unsigned long)pos);
 			return (-1);
 		}
 
-		printf("Value checked array[%zu] = [%d]\n", pos, array[pos]);
+		printf("Value checked array[%lu] = [%d]\n", (unsigned long)pos, array[pos]);
 
 		if (array[pos] == value)
 			return (pos);
